example/django-blog/src/code/requirements.txt
django==3.1.7
django-mdeditor==0.1.18
markdown==3.3.4